+++ src/extension.ts
@@ -202,6 +202,9 @@ export async function activate(context: vscode.ExtensionContext) {
   commandManager.register(new commands.ExportAsPDF(asciidocEngine, context))
   commandManager.register(new commands.PasteImage(context.globalState))
   commandManager.register(new commands.ToggleLockCommand(previewManager))
+  commandManager.register(new commands.ToggleBoldCommand())
+  commandManager.register(new commands.ToggleItalicCommand())
+  commandManager.register(new commands.ToggleMonospaceCommand())
   commandManager.register(new commands.SaveHTML(asciidocEngine))
   commandManager.register(new commands.SaveDocbook(asciidocEngine))
   commandManager.register(